move history table out
